+++ 1-TwoSum/.vs/1-TwoSum/v17/DocumentLayout.json
@@ -1,0 +1,41 @@
+{
+  "Version": 1,
+  "WorkspaceRootPath": "C:\\Users\\Dell\\Desktop\\LeetCode\\1-TwoSum\\",
+  "Documents": [
+    {
+      "AbsoluteMoniker": "D:0:0:{C59BF022-AAE5-4580-978E-153B2A2D38CF}|1-TwoSum\\1-TwoSum.csproj|c:\\users\\dell\\desktop\\leetcode\\1-twosum\\1-twosum\\program.cs||{A6C744A8-0E4A-4FC6-886A-064283054674}",
+      "RelativeMoniker": "D:0:0:{C59BF022-AAE5-4580-978E-153B2A2D38CF}|1-TwoSum\\1-TwoSum.csproj|solutionrelative:1-twosum\\program.cs||{A6C744A8-0E4A-4FC6-886A-064283054674}"
+    }
+  ],
+  "DocumentGroupContainers": [
+    {
+      "Orientation": 0,
+      "VerticalTabListWidth": 256,
+      "DocumentGroups": [
+        {
+          "DockedWidth": 200,
+          "SelectedChildIndex": 1,
+          "Children": [
+            {
+              "$type": "Bookmark",
+              "Name": "ST:1:0:{e8b06f52-6d01-11d2-aa7d-00c04f990343}"
+            },
+            {
+              "$type": "Document",
+              "DocumentIndex": 0,
+              "Title": "Program.cs",
+              "DocumentMoniker": "C:\\Users\\Dell\\Desktop\\LeetCode\\1-TwoSum\\1-TwoSum\\Program.cs",
+              "RelativeDocumentMoniker": "1-TwoSum\\Program.cs",
+              "ToolTip": "C:\\Users\\Dell\\Desktop\\LeetCode\\1-TwoSum\\1-TwoSum\\Program.cs",
+              "RelativeToolTip": "1-TwoSum\\Program.cs",
+              "ViewState": "AgIAABAAAAAAAAAAAAAiwDYAAAAfAAAAAAAAAA==",
+              "Icon": "ae27a6b0-e345-4288-96df-5eaf394ee369.000738|",
+              "WhenOpened": "2025-03-24T18:41:06.629Z",
+              "EditorCaption": ""
+            }
+          ]
+        }
+      ]
+    }
+  ]
+}
