+++ 1-TwoSum/1-TwoSum.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.13.35913.81 d17.13
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "1. Two Sum", "1. Two Sum\1. Two Sum.csproj", "{C75D63D3-6AAE-444B-9BBD-0B2C2D4BE347}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "1-TwoSum", "1-TwoSum\1-TwoSum.csproj", "{C59BF022-AAE5-4580-978E-153B2A2D38CF}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -11,15 +11,15 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{C75D63D3-6AAE-444B-9BBD-0B2C2D4BE347}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C75D63D3-6AAE-444B-9BBD-0B2C2D4BE347}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C75D63D3-6AAE-444B-9BBD-0B2C2D4BE347}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C75D63D3-6AAE-444B-9BBD-0B2C2D4BE347}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C59BF022-AAE5-4580-978E-153B2A2D38CF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C59BF022-AAE5-4580-978E-153B2A2D38CF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C59BF022-AAE5-4580-978E-153B2A2D38CF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C59BF022-AAE5-4580-978E-153B2A2D38CF}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {FFB8313B-5DA9-4DB0-BF25-F1DB2DFF8CEC}
+		SolutionGuid = {7D203318-C259-4255-9815-E92ADE91D46B}
 	EndGlobalSection
 EndGlobal
