+++ 1-TwoSum/1-TwoSum/1-TwoSum.csproj
@@ -4,10 +4,10 @@
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{C75D63D3-6AAE-444B-9BBD-0B2C2D4BE347}</ProjectGuid>
+    <ProjectGuid>{C59BF022-AAE5-4580-978E-153B2A2D38CF}</ProjectGuid>
     <OutputType>Exe</OutputType>
-    <RootNamespace>_1.Two_Sum</RootNamespace>
-    <AssemblyName>_1.Two_Sum</AssemblyName>
+    <RootNamespace>_1_TwoSum</RootNamespace>
+    <AssemblyName>1-TwoSum</AssemblyName>
     <TargetFrameworkVersion>v4.7.2</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <AutoGenerateBindingRedirects>true</AutoGenerateBindingRedirects>
