--- 1. Two Sum/.vs/1. Two Sum/v17/DocumentLayout.json
@@ -1,41 +1,0 @@
-{
-  "Version": 1,
-  "WorkspaceRootPath": "C:\\Users\\Dell\\Desktop\\LeetCode\\1. Two Sum\\",
-  "Documents": [
-    {
-      "AbsoluteMoniker": "D:0:0:{C75D63D3-6AAE-444B-9BBD-0B2C2D4BE347}|1. Two Sum\\1. Two Sum.csproj|c:\\users\\dell\\desktop\\leetcode\\1. two sum\\1. two sum\\program.cs||{A6C744A8-0E4A-4FC6-886A-064283054674}",
-      "RelativeMoniker": "D:0:0:{C75D63D3-6AAE-444B-9BBD-0B2C2D4BE347}|1. Two Sum\\1. Two Sum.csproj|solutionrelative:1. two sum\\program.cs||{A6C744A8-0E4A-4FC6-886A-064283054674}"
-    }
-  ],
-  "DocumentGroupContainers": [
-    {
-      "Orientation": 0,
-      "VerticalTabListWidth": 256,
-      "DocumentGroups": [
-        {
-          "DockedWidth": 200,
-          "SelectedChildIndex": 1,
-          "Children": [
-            {
-              "$type": "Bookmark",
-              "Name": "ST:1:0:{e8b06f52-6d01-11d2-aa7d-00c04f990343}"
-            },
-            {
-              "$type": "Document",
-              "DocumentIndex": 0,
-              "Title": "Program.cs",
-              "DocumentMoniker": "C:\\Users\\Dell\\Desktop\\LeetCode\\1. Two Sum\\1. Two Sum\\Program.cs",
-              "RelativeDocumentMoniker": "1. Two Sum\\Program.cs",
-              "ToolTip": "C:\\Users\\Dell\\Desktop\\LeetCode\\1. Two Sum\\1. Two Sum\\Program.cs*",
-              "RelativeToolTip": "1. Two Sum\\Program.cs*",
-              "ViewState": "AgIAABoAAAAAAAAAAAAEwCAAAAAIAAAAAAAAAA==",
-              "Icon": "ae27a6b0-e345-4288-96df-5eaf394ee369.000738|",
-              "WhenOpened": "2025-03-18T23:51:20.671Z",
-              "EditorCaption": ""
-            }
-          ]
-        }
-      ]
-    }
-  ]
-}
